Customer table functional

+++ Group48_coe528_project/phaseFinal/src/User.java
@@ -8,6 +8,6 @@ public abstract class User {
 
     abstract public String getPassword();
     
-    abstract public int getPoints();
+    abstract public double getPoints();
     
 }
